Move full transaction info adapters into separate files and one package
